Extracted methods from ft helper into an ft builder class - (Task #301)
---
Task #301: Cleanup symmetry reduction code
