+++ de.dlr.sc.virsat.model.extension.fdir.test/src/de/dlr/sc/virsat/model/extension/fdir/converter/DFT2GalileoDFTTest.java
@@ -54,7 +54,7 @@ public class DFT2GalileoDFTTest extends ATestCase {
 		be1.setRepairRate(1);
 		tle.getBasicEvents().add(be1);
 		
-		Gate gate = ftHelper.createGate(tle, FaultTreeNodeType.AND);
+		Gate gate = ftBuilder.createGate(tle, FaultTreeNodeType.AND);
 		
 		Fault fault = new Fault(concept);
 		fault.setName("FAULT");
@@ -65,8 +65,8 @@ public class DFT2GalileoDFTTest extends ATestCase {
 		be2.setHotFailureRate(2);
 		fault.getBasicEvents().add(be2);
 		
-		ftHelper.connect(tle, fault, gate);
-		ftHelper.connect(tle, gate, tle);
+		ftBuilder.connect(tle, fault, gate);
+		ftBuilder.connect(tle, gate, tle);
 		
 		final int TOTAL_NUMBER_OF_BASIC_EVENTS = 2;
 		final int TOTAL_NUMBER_OF_GATES = 3;
