+++ de.dlr.sc.virsat.model.extension.fdir.test/src/de/dlr/sc/virsat/model/extension/fdir/model/FaultTreeTest.java
@@ -27,7 +27,7 @@ import de.dlr.sc.virsat.model.dvlm.structural.StructuralElement;
 import de.dlr.sc.virsat.model.dvlm.structural.StructuralElementInstance;
 import de.dlr.sc.virsat.model.dvlm.structural.StructuralFactory;
 import de.dlr.sc.virsat.model.dvlm.structural.util.StructuralInstantiator;
-import de.dlr.sc.virsat.model.extension.fdir.util.FaultTreeHelper;
+import de.dlr.sc.virsat.model.extension.fdir.util.FaultTreeBuilder;
 
 // *****************************************************************
 // * Import Statements
@@ -62,10 +62,10 @@ public class FaultTreeTest extends AFaultTreeTest {
 		Fault child = new Fault(concept);
 		Fault spare = new Fault(concept);
 		
-		FaultTreeHelper ftHelper = new FaultTreeHelper(concept);
-		ftHelper.connect(root, spareGate, root);
-		ftHelper.connect(root, child, spareGate);
-		ftHelper.connectSpare(root, spare, spareGate);
+		FaultTreeBuilder ftBuilder = new FaultTreeBuilder(concept);
+		ftBuilder.connect(root, spareGate, root);
+		ftBuilder.connect(root, child, spareGate);
+		ftBuilder.connectSpare(root, spare, spareGate);
 		
 		Set<Fault> childSpares = root.getFaultTree().getChildSpares();
 		
@@ -90,10 +90,10 @@ public class FaultTreeTest extends AFaultTreeTest {
 		resource.getContents().add(child.getTypeInstance());
 		resource.getContents().add(grandchild.getTypeInstance());
 		
-		FaultTreeHelper ftHelper = new FaultTreeHelper(concept);
-		ftHelper.connect(root, intermediateNode, root);
-		ftHelper.connect(root, child, intermediateNode);
-		ftHelper.connect(child, grandchild, child);
+		FaultTreeBuilder ftBuilder = new FaultTreeBuilder(concept);
+		ftBuilder.connect(root, intermediateNode, root);
+		ftBuilder.connect(root, child, intermediateNode);
+		ftBuilder.connect(child, grandchild, child);
 		
 		Set<Fault> affectedFaults = child.getFaultTree().getAffectedFaults();
 		
@@ -115,10 +115,10 @@ public class FaultTreeTest extends AFaultTreeTest {
 		
 		sei.getCategoryAssignments().add(root.getTypeInstance());
 		
-		FaultTreeHelper ftHelper = new FaultTreeHelper(concept);
-		ftHelper.connect(root, spareGate, root);
-		ftHelper.connect(root, child, spareGate);
-		ftHelper.connectSpare(root, spare, spareGate);
+		FaultTreeBuilder ftBuilder = new FaultTreeBuilder(concept);
+		ftBuilder.connect(root, spareGate, root);
+		ftBuilder.connect(root, child, spareGate);
+		ftBuilder.connectSpare(root, spare, spareGate);
 		root.getBasicEvents().add(be);
 		be.setRepairAction("Restart");
