+++ de.dlr.sc.virsat.model.extension.fdir.test/src/de/dlr/sc/virsat/model/extension/fdir/recovery/minimizer/FinalStateMinimizerTest.java
@@ -57,7 +57,7 @@ public class FinalStateMinimizerTest extends ATestCase {
 		raHelper.assignInputs(transition01, fault);
 		raHelper.assignInputs(transition11, fault);
 		
-		FaultTreeNode spare = ftHelper.createGate(fault, FaultTreeNodeType.SPARE);
+		FaultTreeNode spare = ftBuilder.createGate(fault, FaultTreeNodeType.SPARE);
 		
 		ClaimAction action = new ClaimAction(concept);
 		action.setClaimSpare(spare);
@@ -98,8 +98,8 @@ public class FinalStateMinimizerTest extends ATestCase {
 		raHelper.assignInputs(transition01_2, fault2);
 		raHelper.assignInputs(transition11_2, fault2);
 		
-		FaultTreeNode spare1 = ftHelper.createGate(fault1, FaultTreeNodeType.SPARE);
-		FaultTreeNode spare2 = ftHelper.createGate(fault2, FaultTreeNodeType.SPARE);
+		FaultTreeNode spare1 = ftBuilder.createGate(fault1, FaultTreeNodeType.SPARE);
+		FaultTreeNode spare2 = ftBuilder.createGate(fault2, FaultTreeNodeType.SPARE);
 		
 		ClaimAction action1 = new ClaimAction(concept);
 		action1.setClaimSpare(spare1);
@@ -144,8 +144,8 @@ public class FinalStateMinimizerTest extends ATestCase {
 		raHelper.assignInputs(transition13, fault1);
 		raHelper.assignInputs(transition23, fault2);
 		
-		FaultTreeNode spare1 = ftHelper.createGate(fault1, FaultTreeNodeType.SPARE);
-		FaultTreeNode spare2 = ftHelper.createGate(fault2, FaultTreeNodeType.SPARE);
+		FaultTreeNode spare1 = ftBuilder.createGate(fault1, FaultTreeNodeType.SPARE);
+		FaultTreeNode spare2 = ftBuilder.createGate(fault2, FaultTreeNodeType.SPARE);
 		
 		ClaimAction action1 = new ClaimAction(concept);
 		action1.setClaimSpare(spare1);
