+++ de.dlr.sc.virsat.model.extension.fdir.test/src/de/dlr/sc/virsat/model/extension/fdir/recovery/minimizer/OrthogonalPartitionRefinementMinimizerTest.java
@@ -83,7 +83,7 @@ public class OrthogonalPartitionRefinementMinimizerTest extends ATestCase {
 		
 		raHelper.assignInputs(transition01, fault1, fault2);
 		
-		FaultTreeNode spare = ftHelper.createGate(fault1, FaultTreeNodeType.SPARE);
+		FaultTreeNode spare = ftBuilder.createGate(fault1, FaultTreeNodeType.SPARE);
 		ClaimAction action = new ClaimAction(concept);
 		action.setClaimSpare(spare);
 		
@@ -119,8 +119,8 @@ public class OrthogonalPartitionRefinementMinimizerTest extends ATestCase {
 		raHelper.assignInputs(transition20, fault1);
 		raHelper.assignInputs(transition21, fault2);
 		
-		FaultTreeNode spare1 = ftHelper.createGate(fault1, FaultTreeNodeType.SPARE);
-		FaultTreeNode spare2 = ftHelper.createGate(fault2, FaultTreeNodeType.SPARE);
+		FaultTreeNode spare1 = ftBuilder.createGate(fault1, FaultTreeNodeType.SPARE);
+		FaultTreeNode spare2 = ftBuilder.createGate(fault2, FaultTreeNodeType.SPARE);
 		
 		ClaimAction action0 = new ClaimAction(concept);
 		action0.setClaimSpare(spare1);
@@ -220,9 +220,9 @@ public class OrthogonalPartitionRefinementMinimizerTest extends ATestCase {
 		raHelper.assignInputs(transition24_2, fault2);
 		raHelper.assignInputs(transition23, fault3);
 		
-		FaultTreeNode spare1 = ftHelper.createGate(fault1, FaultTreeNodeType.SPARE);
-		FaultTreeNode spare2 = ftHelper.createGate(fault2, FaultTreeNodeType.SPARE);
-		FaultTreeNode spare3 = ftHelper.createGate(fault2, FaultTreeNodeType.SPARE);
+		FaultTreeNode spare1 = ftBuilder.createGate(fault1, FaultTreeNodeType.SPARE);
+		FaultTreeNode spare2 = ftBuilder.createGate(fault2, FaultTreeNodeType.SPARE);
+		FaultTreeNode spare3 = ftBuilder.createGate(fault2, FaultTreeNodeType.SPARE);
 		
 		ClaimAction action1 = new ClaimAction(concept);
 		action1.setClaimSpare(spare1);
@@ -276,8 +276,8 @@ public class OrthogonalPartitionRefinementMinimizerTest extends ATestCase {
 		raHelper.assignInputs(transition13, fault1);
 		raHelper.assignInputs(transition23, fault2);
 		                                                                                                                                                                                                    
-		FaultTreeNode spare1 = ftHelper.createGate(fault1, FaultTreeNodeType.SPARE);
-		FaultTreeNode spare2 = ftHelper.createGate(fault2, FaultTreeNodeType.SPARE);
+		FaultTreeNode spare1 = ftBuilder.createGate(fault1, FaultTreeNodeType.SPARE);
+		FaultTreeNode spare2 = ftBuilder.createGate(fault2, FaultTreeNodeType.SPARE);
 		
 		ClaimAction action1 = new ClaimAction(concept);
 		action1.setClaimSpare(spare1);
@@ -330,7 +330,7 @@ public class OrthogonalPartitionRefinementMinimizerTest extends ATestCase {
 		raHelper.assignInputs(transition24, fault1);
 		raHelper.assignInputs(transition34, fault1);
 		
-		FaultTreeNode spare1 = ftHelper.createGate(fault1, FaultTreeNodeType.SPARE);
+		FaultTreeNode spare1 = ftBuilder.createGate(fault1, FaultTreeNodeType.SPARE);
 		
 		ClaimAction action1 = new ClaimAction(concept);
 		action1.setClaimSpare(spare1);
